Move code from BackgroundScanner into the CronJob, extract DB query into a separate class
