+++ appinfo/application.php
@@ -18,7 +18,7 @@ use OCA\Files_Antivirus\AvirWrapper;
 use OCA\Files_Antivirus\Controller\RuleController;
 use OCA\Files_Antivirus\Controller\SettingsController;
 use OCA\Files_Antivirus\Db\RuleMapper;
-use OCA\Files_Antivirus\BackgroundScanner;
+use OCA\Files_Antivirus\Db\FileCollection;
 use OCA\Files_Antivirus\RequestHelper;
 use OCA\Files_Antivirus\ScannerFactory;
 use OCP\AppFramework\App;
@@ -69,16 +69,12 @@ class Application extends App {
 				);
 			}
 		);
-		
+
 		$container->registerService(
-			'BackgroundScanner',
+			'FileCollection',
 			function ($c) {
-				return new BackgroundScanner(
-					$c->query('ScannerFactory'),
-					$c->query('L10N'),
-					$c->query('AppConfig'),
-					$c->getServer()->getRootFolder(),
-					$c->getServer()->getUserSession()
+				return new FileCollection(
+					$c->query('ServerContainer')->getDatabaseConnection()
 				);
 			}
 		);
